pyproject.toml
[project]
name = "currency-agent"
version = "0.1.0"
description = "货币转换和汇率查询 Agent"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk==0.2.10",
    "requests",
    "aiohttp",
    "uvicorn"
] 